refactor: split counter logic and add folder structure
- split components into folders: components, containers, hooks
- split counter logic into useCounter hook
