+++ src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Counter from './Counter';
+import Counter from './containers/Counter';
 
 function App() {
   return (
